refactor(textarea): move to script setup
