+++ src/packages/__VUE/textarea/doc.md
@@ -59,6 +59,14 @@ app.use(Textarea)
 | focus | 聚焦时触发 | `event` |
 | blur | 失焦时触发 | `value,event` |
 
+### 类型定义 version
+
+组件导出以下类型定义：
+
+```js
+import type { TextareaProps, TextareaInstance } from '@nutui/nutui';
+```
+
 ## 主题定制
 
 ### 样式变量
